base repository

+++ src/Repositories/BaseRepository.php
@@ -108,15 +108,17 @@ class BaseRepository
      *
      * return       int     更新資料的自動編號
      */
-    public function replicateUpdate($id, $manually = false)
+    public function replicateUpdate($id, $manually = false, $use_replicate = true)
     {
         $q = $this->model->find($id);
         if (!is_null($q)) {
-            $newVersion = $q->replicate();
-            $newVersion->old_version = 1;
-            $newVersion->origin_id = $id;
-            $newVersion->save();
-            $newVersion->delete();
+            if ($use_replicate) {
+                $newVersion = $q->replicate();
+                $newVersion->old_version = 1;
+                $newVersion->origin_id = $id;
+                $newVersion->save();
+                $newVersion->delete();
+            }
             return $this->executeUpdate($id, $this->makeUpdateData($manually));
         } else {
             dd('BaseRepository::replicateUpdate $id error');
